+++ src/layers/cip/core/feature.js
@@ -1,9 +1,10 @@
 'use strict';
 
 class CIPFeature {
-  constructor(code, name) {
+  constructor(code, name, classCode) {
     this.code = code;
     this.name = name;
+    this.classCode = classCode;
   }
 }