+++ src/layers/cip/core/objects/TCPIPInterface.js
@@ -2,9 +2,9 @@
 
 const CIPMetaObject = require('../object');
 const CIPAttribute = require('../attribute');
-const CIPFeatureGroup = require('../featuregroup');
 const { ClassCodes } = require('../constants');
 const { DataType } = require('../datatypes');
+
 
 const ClassAttribute = Object.freeze({});
 
@@ -69,10 +69,6 @@ const InstanceAttribute = Object.freeze({
 });
 
 
-const ClassAttributeGroup = new CIPFeatureGroup(Object.values(ClassAttribute))
-const InstanceAttributeGroup = new CIPFeatureGroup(Object.values(InstanceAttribute));
-
-
 const GetAttributesAllInstanceAttributes = Object.freeze([
   InstanceAttribute.Status,
   InstanceAttribute.ConfigurationCapability,
@@ -84,8 +80,8 @@ const GetAttributesAllInstanceAttributes = Object.freeze([
 
 
 const CIPObject = CIPMetaObject(ClassCodes.TCPIPInterface, {
-  ClassAttributeGroup,
-  InstanceAttributeGroup,
+  ClassAttributes: ClassAttribute,
+  InstanceAttributes: InstanceAttribute,
   GetAttributesAllInstanceAttributes
 });