+++ src/utils.js
@@ -167,6 +167,29 @@ function CallbackPromise(callback, func, timeout) {
 }
 
 
+// class Deferred {
+//   constructor() {
+//     this.promise = new Promise((resolve, reject) => {
+//       this.reject = reject
+//       this.resolve = resolve
+//     });
+//   }
+// }
+
+function deferred() {
+  let _resolve, _reject;
+  const promise = new Promise((resolve, reject) => {
+    _resolve = resolve;
+    _reject = reject;
+  });
+
+  promise.resolve = _resolve;
+  promise.reject = _reject;
+
+  return promise;
+}
+
+
 module.exports = {
   getBits,
   getBit,
@@ -176,5 +199,7 @@ module.exports = {
   once,
   InvertKeyValues,
   CallbackPromise,
-  InfoError
+  InfoError,
+  // Deferred
+  deferred
 };